test function exchange server
