+++ .env
@@ -1,1 +1,1 @@
-VITE_API_URL=http://localhost:8000
+VITE_API_URL=https://caritas.rassokha.pro/api/v1
