+++ src/lib/api/endpoints.ts
@@ -1,4 +1,4 @@
 export const API_ENDPOINTS = {
-  upload: 'assessments/upload',
-  results: '/assessments/{request_id}',
+  upload: '/assessments/upload',
+  results: '/assessments/',
 };
