+++ src/pages/Home/HomePage.tsx
@@ -6,10 +6,15 @@ import { DatePicker } from '../Home/components/date-picker';
 import { useEffect, useState } from 'react';
 import { InstructionPopup } from '../Home/components/InstructionPopup';
 import { MainBlockForm } from '../Home/components/MainBlockForm';
+import { Loader } from '../Home/components/loader';
+import { ValidErrorPopup } from '../Home/components/ValidErrorPopup';
 
 function HomePage() {
   const [selectDate, setSelectDate] = useState('');
   const [isOpenPopup, setIsOpenPopup] = useState(false);
+  const [openLoader, setOpenLoader] = useState(false);
+  const [isErrorPopupOpen, setIsErrorPopupOpen] = useState(false);
+  const [errorMessage, setErrorMessage] = useState<string[]>([]);
 
   const handleDate = (newDate: string) => {
     setSelectDate(newDate);
@@ -27,29 +32,63 @@ function HomePage() {
     }
   }, []);
 
+  const toggleLoader = () => {
+    setOpenLoader(prev => !prev);
+  };
+
+  const openErrorPopup = (errors: string[]) => {
+    setErrorMessage(errors);
+    setIsErrorPopupOpen(true);
+  };
+
+  const closeErrorPopup = () => {
+    setIsErrorPopupOpen(false);
+    setErrorMessage([]);
+  };
+
   return (
-    <MainWrapper>
-      <header className={styles.header}>
-        <div className={styles.logoBlock}>
-          <Logo />
-        </div>
-        <div className={styles.titleBlock}>
-          <TitleSection />
-        </div>
-        <div className={styles.dateBlock}>
-          <DatePicker
-            value={selectDate}
-            onChange={handleDate}
-            label={'Дата заполнения'}
-            required={true}
+    <section>
+      <div
+        className={
+          openLoader
+            ? `${styles.pageContent} ${styles.pageContentWithLoader}`
+            : styles.pageContent
+        }
+      >
+        <MainWrapper>
+          <header className={styles.header}>
+            <div className={styles.logoBlock}>
+              <Logo />
+            </div>
+            <div className={styles.titleBlock}>
+              <TitleSection />
+            </div>
+            <div className={styles.dateBlock}>
+              <DatePicker
+                value={selectDate}
+                onChange={handleDate}
+                label={'Дата заполнения'}
+                required={true}
+              />
+            </div>
+          </header>
+          <main className={styles.main}>
+            <MainBlockForm
+              openPopup={handleTargetPopup}
+              openLoader={toggleLoader}
+              openErrorPopup={openErrorPopup}
+            />
+          </main>
+          <InstructionPopup isOpen={isOpenPopup} doClose={handleTargetPopup} />
+          <ValidErrorPopup
+            isOpen={isErrorPopupOpen}
+            doClose={closeErrorPopup}
+            errors={errorMessage}
           />
-        </div>
-      </header>
-      <main className={styles.main}>
-        <MainBlockForm openPopup={handleTargetPopup} />
-      </main>
-      <InstructionPopup isOpen={isOpenPopup} doClose={handleTargetPopup} />
-    </MainWrapper>
+        </MainWrapper>
+      </div>
+      {openLoader && <Loader />}
+    </section>
   );
 }
