+++ src/pages/Home/styles.module.css
@@ -36,6 +36,13 @@
   -o-transform: translateY(-4px);
 }
 
+.pageContentWithLoader {
+  position: relative;
+  z-index: 1;
+  opacity: 0.6;
+  filter: blur(0.2rem);
+}
+
 @media screen and (max-width: 1199px) {
   .titleBlock {
     height: 100%;
